dec02 task. Required adding multiline support for the base classes.
Add support for multiline input or output from the test files. This
resulted in quite the rewrite of most i/o parts of the main framework.

When solving dec02_2 use the fact that the nominator much be bigger then
the denominator for the division to happen. So sort the lists, then loop
until we hit a pair where this does not add up.