Change handling of test input and result.
Allow tasks to set if they want integer input and honor this for all
input data.
Select if we want a list input or a single line by implenting run_list()
or run() respectively.
Return list (of str or int), str or int and make the result comparison
handle it.
This also removes main(), use runtests() instead.