Major refacttor of skeleton code
Use a base class in skel as we will almost always use a common parse
method between parts. Also allow this to return something passed to run
or use class vaiables to relay more advanced data.

Better handling of all sorts of test inputs and results. Use a table for
rendering output when lists are involved or simply diff them. Never echo
input as this just messes up the output.